Refactor code for data_processing and training NN
Put all scripts in 1 file called nn_script
Combine labeling and feature_extraction into 1 file called data_processing
